Changed formulae.py file to begin splitting backward and predict steps
Did this split because there can potentially be more averaging (in time
and freq) on the backward step (where we grid the visibilities) than is
possible on the forward step (where we predict what the visibilities
should look like according to the sky model).
This is a work in progress, affecting the girding rate and the
convolution function calculation rate.