Further development of kernel re-use and possible on the fly calc
kernel re-use following verbal discussion with Paul and Anna. (to be
checked)
optional “on the fly” kernel calculations checking potential savings if
we don’t over sample the kernel but use a new one for each and every
visibility….work in progress. (The on the fly parameter is 0-1 but
should be True/False.)