Prevent Nfacet search from stopping too soon
Prevent a local minimum at Nfacet=1 from forcing the search over
Nfacets stopping too soon - Nfacet=2 is artificially high because of
the overlap fraction cost introduced with N>1 facets. This change
forces all values of Nfacet from 1-5 to be searched (and the search
still continues if a minimum is still not found).