Use lambda for reasoning about baseline bins
The idea here is that this might allow us to move the baseline loop
into equations.py, see sum_bl_bins. This makes baseline dependence
explit, at the cost of some noise for extra lambda statements. Sadly
creating the sum term with sympy also turns out to be a bit of a
bottleneck, with the bcount / bmax substitution as the main culprit.