Heavy refactoring of api_ipython
Pulls most of the configuration checking that was duplicated all over
the place into PipelineConfig. We can also generalise pretty much all
evaluation using a slightly more flexible variant of _compute_results.

This also changes the approach of parameter sweeps a bit - instead of
setting a variable and re-running the imaging equations, we introduce
the adjustments in calc_tel_params. This is the only way that we can
get the right bin distribution for the new binning method.

At this point, most functionality should work again. Let's hope I didn't
break anything in a non-obvious way...