Region zipping
For reasoning about edges we need a way to look at all region box pairs
on both sides. This is what a "zip" is supposed to do: Allow us to sum a
property over all such pairs.

The number of cases here is fairly large, as we can have domains present
or missing on either side, plus enumerated vs not enumerated. Handling
all this consistently isn't quite obvious, so I added a bunch of test
cases to make sure we got it right.