Some graph adjustments
Hide the totals by default, use proper order for stacked bars, actually
implement GRAPH_ROWS. We now use the same names for the table as for the
graph, and skip drawing the graph completely if rows are missing or in
the wrong order. Crude, but this should have the best shot at preventing
misunderstandings.