Data flow
The goal here is to describe the full pipeline data flow in enough
detail to automate certain analyses on them. As the full graph is
expected to be very, very large, this means we need to be able to
reason about large sets of nodes at the same time.

The main idea here is that region boxes span up the whole graph, but
we actually only enumerate the node "classes" that have different
parameters. For example, this means that for most nodes right now we
only need to enumerate baseline bins in order to calculate all numbers
that are interesting about the graph.

Plan here is that this should allow us to automatically reason about
things like in-/out-degrees as well. This should eventually allow us
to exactly determine the total load on the data flow system.