Introduce true full-FoV visibility rate
This replaces Nvis_predict_no_averaging with Nvis, which only does
averaging to the point that we retain enough information to produce
correct facet visibilities (which might then get averaged more
aggressively).

Chose the simple name, as this is really the base visibility rate, and
especially has basically nothing to do with predict any more. Not
completely convinced that it shouldn't be BL-dependent, but this was
the idea before, so let's stick with it.