Symbolify product equations
This allows us to quickly generate a somewhat sane product equations
list. Basic idea is that we simply replace all symbols in the
parameter container by symbols at strategic points in
equations.py. This gives us a per-definition complete view of the
whole situation.

Behind the scenes, this also adds a bit of ad-hoc term optimisation in
_sum_bl_bins. Might speed up things a little, who knows.