Dataflow refactoring following Tim's changes
This also changes the entire syntax for how we're definition products.
The idea here is basically to make a proper pattern out of the common
pattern "task multiplier * task rate": We still declare a bunch of
rates (such as "Rflop" and/or "Rout"). Yet we now also have a special "N"
parameter, which scales all of these rates and defines task granularity
at the same time.