Symbolic dataflow region properties
This allows us to define region properties that depend on the region
identity without actually enumerating individual regions. Instead, using
sympy formulas we can leave the region index as a symbol and resolve it
later (or not at all, to return the formula).

The details are slightly tricky here - we are probably over-using duct-
typing quite a bit. Not quite sure how to fix that right now.