Replace Lambda by BLDep
BLDep acts exactly like a numpy Lambda, but also documents exactly what
the lambda means. This is cleaner than the solution we had before.

The most notable change in practice is that this allows us to merge
"_set_product" and "_set_product_blsum" by pushing the baseline dependence
to the caller. They can then choose to use "BLDep" or "blsum" dependending
on whether the value in question is actually meant to be summed or not.

This fixes a rather subtle error with the output size calculation of
gridding.