Add ability to query box multipliers
Under the hood, this changes semantics of INDEX a bit - now it refers
to the *first* index of the region set. Before, we assumed it should be
*the* index of the current box, which would only be available if the
region set had size 1 (= enumerated).

Not sure. COUNT was already magical, so now INDEX is as well. Maybe tag
them as such in some way?