Fix breakage due to multiplier change
This was a actually a bit subtle, causing inconsistencies in the data
flow output. Should now be well covered by test cases.

The whole multiplier business will need test cases too at some point.
Still not quite sure what the semantics are supposed to be in the
first place though.