Do not sum w cache by baseline
Historically, the formula has actually never included a baseline sum
(my mistake), but it was summed over baseline bins (previous mistake).
Realistically, the w cache as described here would only require to go
to the *maximum* size, and be evaluated up to the *maximum* frequency
channel count. Everything else is simple subgrids/sets of this.