Formula overview improvements
Allow batch-generation of formulas for multiple configurations. Also
include values determined by optimisation into the constant table. This
obviously involves actually running the optimisation (which is meaningless
for symbolic views), but it still makes the view more consistent.

Also switch collect_free_symbols to return a set of strings (instead of
Symbols), which resolves a bunch of weirdness. Finally generate headings
as Markdown and use "aligned" instead of "align" to allow LaTeX exports
to work.