Merge kernel generation
Merging kernel generation into gridding/degridding is the more honest
way of interpreting this - there is simply no way kernel generation
as formulated can be seen as a separate product.

Also slightly adjust pipelines to generate, and link the resulting PDFs.
Hopefully this allows us to embed them for the Jenkins export.