Split CSV generation/comparison out of imaging performance notebooks
This means we need to move the "available" lists into the parameter
definitions so we can share them. This seems like a natural choice to me.

The new "export" notebook now has a few more features, such as chain-
comparing pipelines and HPSOs results based on Git history. It will
especially fail if the latest versions checked into Git do not match
current results.

Furthermore, there are a few small changes in the imaging notebook as well.
Most of it is layout, but now we also use "interact" more often. The code
is quick enough that we can afford a bit more snappiness, and this way
Jenkins will end up running it as well.