Symbol renamings and comment updates
An attempt at making things more consistent with the parametric model
document. This removes all comments that are likely out of date (e.g.
refer to PDR05) or refer to past reviews. I added coarse references to
the delta-PDR sections, that might actually be maintainable.

Furthermore, a bunch of symbols have changed name in order to be more
consistent and generate better formulas:

 * nbaselines -> Nbl
 * minimum_channels -> Nf_min
 * minimum_channels_gran -> Nf_min_gran
 * number_taylor_terms -> Ntt
 * facet_overlap_fraq -> r_facet_base
 * using_facet_overlap_frac -> r_facet
 * nr_baselines -> the abyss, where it belongs
 * NAProducts -> disabled until we actually implement it (any second now!)
 * Total_fov -> Theta_fov_total
 * Ntaylor -> Ntt