Flow transfer splits, reorganisation and cleanups
The aim with making "split" Flows at a given granularity is to allow us
to model flows that send only "relevant" data downstream. This is
actually relevant to realistic pipeline generation, hence it coming
up now. Currently we set the "worker" Flow to have zero output. This is
probably the right move right now, but might eventually need special
cases in generation.

Also refactor merge so it properly calculates the output data using the
transfer formula. This needs some plumbing for the case where a domain
vanishes that has properties that the formula uses (e.g. baseline length).
This is not quite trivial, so add appropriate test coverage as well.