Calculate time coalescing amount using facet FoV
The old method used the full FoV value, scaled by Nfacet/(1+r_facet).
That's *almost* the same thing, as the FoV itself scales with that
factor and the FoV factors in linearly. However, in between we round
and apply bounds to the number, which is not appropriate.

This yields less averaging on long baselines, and overall less jitter due
to less coarse rounding. Interestingly enough, these two effects seem to
pretty much exactly cancel each other out in terms of cost.

# Conflicts:
#	iPython/equations.py