Nerf domains directly instead of messing with resulting multipliers
This is essentially about only considering a part of the full pipeline,
so it is probably the right place to have the pipeline handle it. Makes
domain declarations more complicated, but live becomes easier for the
generator. Added some (weak) test cases to it so it has a chance to
retain stability.