Overhaul of domain-dependent expressions
Replaces some more lambda magic by an "explicit" solution: A DomainExpr
is an expression that depends on properties of some domain. Primitive
domain-dependent expressions can be constructed using __call__ of domains,
so e.g. "baseline('size')" is now a baseline size expression. Can be
combined using a few operators. Evaluate using the usual the() / sum() /
max() functions.

Only downside here is that it can't handle cases well where we have two
region boxes that we care about, such as with edges. Not quite sure
what to do about that yet.