Generalise parameter optimisation
Before, optimisation for Tsnap / Nfacet was hard-coded into the
parametric model. This isn't really necessary - we can just look at
the free variables in the to-optimised expression, and adjust those
back.

This also changes "calc_tel_params" to return optimised parameters by
default - this is what we want in 99% of the cases. This saves a bunch
of boilerplate code in a few places.