New attempt at scheduling
There's a couple of differences here to Francois' scheduler, both
conceptual and technical. Most important is likely that we use a level
trace that always returns to zero in order to reason about
scheduling. This makes it easier to prevent "leaks" from happening,
and especially means we can guarantee that every task that fits into
the capacities will be scheduled eventually.

The way this works is that for "edge costs" (e.g. buffer objects) we
first reserve cost just to the end of the current task, and any future
dependent task needs to "extend" the edge. If this fails, we know that
we need to re-schedule the original task past the point where the edge
extension failed (i.e. the buffer was too full). This leads to a
comparatively easy back-tracking scheme that eventually places every
task.

Result is that Low is *barely* realistic, and Mid needs quite a bit
more buffer and especially Delivery capacity. Will need to check further.