Allow converting csv into pre-stripped dictionaries for fast lookup
Looking up CSV values is pretty slow due to the fact that we need to
strip the row and column names every time. If we do this up-front, we
can actually employ dictionary lookups.

This is a bit hacky - we don't want to make this the default, because
in some cases we actually care about row and column order. So here we
just introduce a function that converts the csv data into a "stripped"
representation, and adjust lookup_csv so it works with both.