Introduce sensitivity analysis
The idea here is to look into the interaction between observatory
schedules and efficiency. Clearly a really bad schedule can cause SDP
to slow the entire telescope down - yet we cannot assume that the
observatory is going to tailor its entire schedule after the needs of
SDP.

Therefore this analysis walks through a big number of possible
schedules, and calculates the percentiles on efficiency changes
depending on capacity. This should hopefully give us a way to take a
bit of magic out of system sizing.

Also refactored a bit - moved quite a bit of code around, both inside
the notebook and from the notebook into the modules.