Clean up Git history handling
Tinker a bit with how we prioritise the CSV to compare against so that
it hopefully becomes more stable against merge confusion. It should
not go hierarchically, then by date (--date-order).

Also clean up the export notebook, as a lot of code had actually moved
to reports.py. Finally - give more helpful messages if this happens in
the future.