Support for re-scheduling
This models the need to re-schedule due to unforeseen
circumstances (such as a large-scale hardware or software
failure). Consequently, this starts from a schedule appropriate to the
previous assumptions (i.e. full capacity), and returns

1. a new schedule, where from the designated point forward the task
   schedule is updated to honour the new constraints

2. "failed" tasks / resources, which is resources irrecoverably lost
   due to the unplanned outtage. Note that for the purpose of this
   model, this goes all the way to re-observing data if it turns out
   that the input buffer can't hold previously observed data any more.

There's a lot of alternative failure scenarios we might want to
consider, but this feels like a pretty honest way of going about it
at this abstraction level.