Update top-level documentation
Most of the text copied from the PM's user guide document doesn't
really make sense as documentation. Idea is to make this the
jumping-off point to the notebooks - will likely need to work on
things quite a bit further, but this should be a start.

This especially attempts to set up GitLab CI to integrate Sphinx and
iPython documentation appropriately (as well as coverage, while we're
at it).