Introduce linked capacities and cost analysis
We cannot really consider buffer capacity independently of its
throughput - so here we lock the ratio between them using data from
the cost model.

Furthermore, it is quite hard to put into perspective how "bad" it is
to lose efficiency compared with potential savings on hardware. To
make this easier, we introduce a way to contrast efficiency loss (in
terms of cost of the entire project) compared to the hardware
cost (again using approximate cost model data).

All of this adds a bunch of complexity, so move it to its own module.